Show the optimized and non-optimized query plans

These functions are available for LazyFrames only. describe_plan() shows the query plan as-is, without any optimization done by Polars. This is not the query performed. Before running the query, Polars applies a list of optimizations (such as filtering data before sorting it). The actual query plan ran by Polars is shown by describe_optimized_plan(). Note that the queries are read from bottom to top.

Usage

describe_plan(.data)

describe_optimized_plan(.data)

Arguments

.data

A Polars LazyFrame

Examples

query <- mtcars |>
  as_polars_lf() |>
  arrange(drat) |>
  filter(cyl == 3) |>
  select(mpg)

describe_plan(query)
#>  SELECT [col("mpg")] FROM
#>   FILTER [(col("cyl")) == (3.0)] FROM
#>     SORT BY [col("drat")]
#>       DF ["mpg", "cyl", "disp", "hp"]; PROJECT */11 COLUMNS; SELECTION: None
#> $ok
#> NULL
#> 
#> $err
#> NULL
#> 
#> attr(,"class")
#> [1] "extendr_result"

describe_optimized_plan(query)
#> simple π 1/3 ["mpg"]
#>   SORT BY [col("drat")]
#>     DF ["mpg", "cyl", "disp", "hp"]; PROJECT 3/11 COLUMNS; SELECTION: [(col("cyl")) == (3.0)]