Create, modify, and delete columns

This creates new columns that are functions of existing variables. It can also modify (if the name is the same as an existing column) and delete columns (by setting their value to NULL).

Usage

pl_mutate(.data, ...)

Arguments

.data

A Polars Data/LazyFrame

...

Name-value pairs. The name gives the name of the column in the output. The value can be:

A vector the same length as the current group (or the whole data frame if ungrouped).

NULL, to remove the column.

Details

A lot of functions available in base R (cos, sin, multiplying, etc.) or in other packages (dplyr::lag(), etc.) are implemented in an efficient way in Polars. These functions will be automatically translated to Polars syntax under the hood so that you can continue using the classic R syntax and functions.

If a Polars built-in replacement doesn't exist (for example for custom functions), the R function will be passed to map() in the Polars workflow. Note that this is slower than using functions that can be translated to Polars syntax.

Examples

pl_iris <- polars::pl$DataFrame(iris)

# classic operation
pl_mutate(pl_iris, x = Sepal.Width + Sepal.Length)
#> shape: (150, 6)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┬─────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   ┆ x   │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       ┆ --- │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       ┆ f64 │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╪═════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 8.6 │
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 7.9 │
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    ┆ 7.9 │
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    ┆ 7.7 │
#> │ …            ┆ …           ┆ …            ┆ …           ┆ …         ┆ …   │
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica ┆ 8.8 │
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica ┆ 9.5 │
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica ┆ 9.6 │
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica ┆ 8.9 │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┴─────┘

# logical operation
pl_mutate(pl_iris, x = Sepal.Width > Sepal.Length & Petal.Width > Petal.Length)
#> shape: (150, 6)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┬───────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   ┆ x     │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       ┆ ---   │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       ┆ bool  │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╪═══════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ false │
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ false │
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    ┆ false │
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    ┆ false │
#> │ …            ┆ …           ┆ …            ┆ …           ┆ …         ┆ …     │
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica ┆ false │
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica ┆ false │
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica ┆ false │
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica ┆ false │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┴───────┘

# overwrite existing variable
pl_mutate(pl_iris, Sepal.Width = Sepal.Width*2)
#> shape: (150, 5)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╡
#> │ 5.1          ┆ 7.0         ┆ 1.4          ┆ 0.2         ┆ setosa    │
#> │ 4.9          ┆ 6.0         ┆ 1.4          ┆ 0.2         ┆ setosa    │
#> │ 4.7          ┆ 6.4         ┆ 1.3          ┆ 0.2         ┆ setosa    │
#> │ 4.6          ┆ 6.2         ┆ 1.5          ┆ 0.2         ┆ setosa    │
#> │ …            ┆ …           ┆ …            ┆ …           ┆ …         │
#> │ 6.3          ┆ 5.0         ┆ 5.0          ┆ 1.9         ┆ virginica │
#> │ 6.5          ┆ 6.0         ┆ 5.2          ┆ 2.0         ┆ virginica │
#> │ 6.2          ┆ 6.8         ┆ 5.4          ┆ 2.3         ┆ virginica │
#> │ 5.9          ┆ 6.0         ┆ 5.1          ┆ 1.8         ┆ virginica │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┘

# grouped computation
pl_iris |>
  pl_group_by(Species) |>
  pl_mutate(
    foo = mean(Sepal.Length)
  )
#> shape: (150, 6)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┬───────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   ┆ foo   │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       ┆ ---   │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       ┆ f64   │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╪═══════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 5.006 │
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 5.006 │
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    ┆ 5.006 │
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    ┆ 5.006 │
#> │ …            ┆ …           ┆ …            ┆ …           ┆ …         ┆ …     │
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica ┆ 6.588 │
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica ┆ 6.588 │
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica ┆ 6.588 │
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica ┆ 6.588 │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┴───────┘

# across() is available
pl_iris |>
  pl_mutate(
    across(.cols = contains("Sepal"), .fns = mean, .names = "{.fn}_of_{.col}")
  )
#> shape: (150, 7)
#> ┌────────────┬─────────────┬───────────────┬─────────────┬───────────┬──────────────┬──────────────┐
#> │ Sepal.Leng ┆ Sepal.Width ┆ Petal.Length  ┆ Petal.Width ┆ Species   ┆ mean_of_Sepa ┆ mean_of_Sepa │
#> │ th         ┆ ---         ┆ ---           ┆ ---         ┆ ---       ┆ l.Length     ┆ l.Width      │
#> │ ---        ┆ f64         ┆ f64           ┆ f64         ┆ cat       ┆ ---          ┆ ---          │
#> │ f64        ┆             ┆               ┆             ┆           ┆ f64          ┆ f64          │
#> ╞════════════╪═════════════╪═══════════════╪═════════════╪═══════════╪══════════════╪══════════════╡
#> │ 5.1        ┆ 3.5         ┆ 1.4           ┆ 0.2         ┆ setosa    ┆ 5.843333     ┆ 3.057333     │
#> │ 4.9        ┆ 3.0         ┆ 1.4           ┆ 0.2         ┆ setosa    ┆ 5.843333     ┆ 3.057333     │
#> │ 4.7        ┆ 3.2         ┆ 1.3           ┆ 0.2         ┆ setosa    ┆ 5.843333     ┆ 3.057333     │
#> │ 4.6        ┆ 3.1         ┆ 1.5           ┆ 0.2         ┆ setosa    ┆ 5.843333     ┆ 3.057333     │
#> │ …          ┆ …           ┆ …             ┆ …           ┆ …         ┆ …            ┆ …            │
#> │ 6.3        ┆ 2.5         ┆ 5.0           ┆ 1.9         ┆ virginica ┆ 5.843333     ┆ 3.057333     │
#> │ 6.5        ┆ 3.0         ┆ 5.2           ┆ 2.0         ┆ virginica ┆ 5.843333     ┆ 3.057333     │
#> │ 6.2        ┆ 3.4         ┆ 5.4           ┆ 2.3         ┆ virginica ┆ 5.843333     ┆ 3.057333     │
#> │ 5.9        ┆ 3.0         ┆ 5.1           ┆ 1.8         ┆ virginica ┆ 5.843333     ┆ 3.057333     │
#> └────────────┴─────────────┴───────────────┴─────────────┴───────────┴──────────────┴──────────────┘
# It can receive several types of functions:
pl_iris |>
  pl_mutate(
    across(
      .cols = contains("Sepal"),
      .fns = list(mean = mean, sd = ~ sd(.x)),
      .names = "{.fn}_of_{.col}"
    )
  )
#> shape: (150, 9)
#> ┌────────────┬───────────┬────────────┬───────────┬───┬────────────┬────────────┬────────────┬────────────┐
#> │ Sepal.Leng ┆ Sepal.Wid ┆ Petal.Leng ┆ Petal.Wid ┆ … ┆ mean_of_Se ┆ sd_of_Sepa ┆ mean_of_Se ┆ sd_of_Sepa │
#> │ th         ┆ th        ┆ th         ┆ th        ┆   ┆ pal.Length ┆ l.Length   ┆ pal.Width  ┆ l.Width    │
#> │ ---        ┆ ---       ┆ ---        ┆ ---       ┆   ┆ ---        ┆ ---        ┆ ---        ┆ ---        │
#> │ f64        ┆ f64       ┆ f64        ┆ f64       ┆   ┆ f64        ┆ f64        ┆ f64        ┆ f64        │
#> ╞════════════╪═══════════╪════════════╪═══════════╪═══╪════════════╪════════════╪════════════╪════════════╡
#> │ 5.1        ┆ 3.5       ┆ 1.4        ┆ 0.2       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ 4.9        ┆ 3.0       ┆ 1.4        ┆ 0.2       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ 4.7        ┆ 3.2       ┆ 1.3        ┆ 0.2       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ 4.6        ┆ 3.1       ┆ 1.5        ┆ 0.2       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ …          ┆ …         ┆ …          ┆ …         ┆ … ┆ …          ┆ …          ┆ …          ┆ …          │
#> │ 6.3        ┆ 2.5       ┆ 5.0        ┆ 1.9       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ 6.5        ┆ 3.0       ┆ 5.2        ┆ 2.0       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ 6.2        ┆ 3.4       ┆ 5.4        ┆ 2.3       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> │ 5.9        ┆ 3.0       ┆ 5.1        ┆ 1.8       ┆ … ┆ 5.843333   ┆ 0.828066   ┆ 3.057333   ┆ 0.435866   │
#> └────────────┴───────────┴────────────┴───────────┴───┴────────────┴────────────┴────────────┴────────────┘

# Embracing an external variable works
some_value <- 1
pl_mutate(pl_iris, x = {{ some_value }})
#> shape: (150, 6)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┬─────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   ┆ x   │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       ┆ --- │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       ┆ f64 │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╪═════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 1.0 │
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 1.0 │
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    ┆ 1.0 │
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    ┆ 1.0 │
#> │ …            ┆ …           ┆ …            ┆ …           ┆ …         ┆ …   │
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica ┆ 1.0 │
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica ┆ 1.0 │
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica ┆ 1.0 │
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica ┆ 1.0 │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┴─────┘