Pivot Data/LazyFrame from wide to long

Pivot Data/LazyFrame from wide to long

Usage

pl_pivot_longer(data, cols, names_to = "name", values_to = "value")

Arguments

data

A Polars Data/LazyFrame

cols

Columns to pivot into longer format. Can be anything accepted by dplyr::select().

names_to

The (quoted) name of the column that will contain the column names specified by cols.

values_to

A string specifying the name of the column to create from the data stored in cell values.

Examples

pl_relig_income <- polars::pl$DataFrame(relig_income)
pl_relig_income
#> shape: (18, 11)
#> ┌─────────────────┬───────┬─────────┬─────────┬───┬──────────┬───────────┬───────┬──────────────┐
#> │ religion        ┆ <$10k ┆ $10-20k ┆ $20-30k ┆ … ┆ $75-100k ┆ $100-150k ┆ >150k ┆ Don't        │
#> │ ---             ┆ ---   ┆ ---     ┆ ---     ┆   ┆ ---      ┆ ---       ┆ ---   ┆ know/refused │
#> │ str             ┆ f64   ┆ f64     ┆ f64     ┆   ┆ f64      ┆ f64       ┆ f64   ┆ ---          │
#> │                 ┆       ┆         ┆         ┆   ┆          ┆           ┆       ┆ f64          │
#> ╞═════════════════╪═══════╪═════════╪═════════╪═══╪══════════╪═══════════╪═══════╪══════════════╡
#> │ Agnostic        ┆ 27.0  ┆ 34.0    ┆ 60.0    ┆ … ┆ 122.0    ┆ 109.0     ┆ 84.0  ┆ 96.0         │
#> │ Atheist         ┆ 12.0  ┆ 27.0    ┆ 37.0    ┆ … ┆ 73.0     ┆ 59.0      ┆ 74.0  ┆ 76.0         │
#> │ Buddhist        ┆ 27.0  ┆ 21.0    ┆ 30.0    ┆ … ┆ 62.0     ┆ 39.0      ┆ 53.0  ┆ 54.0         │
#> │ Catholic        ┆ 418.0 ┆ 617.0   ┆ 732.0   ┆ … ┆ 949.0    ┆ 792.0     ┆ 633.0 ┆ 1489.0       │
#> │ …               ┆ …     ┆ …       ┆ …       ┆ … ┆ …        ┆ …         ┆ …     ┆ …            │
#> │ Other Christian ┆ 9.0   ┆ 7.0     ┆ 11.0    ┆ … ┆ 18.0     ┆ 14.0      ┆ 12.0  ┆ 18.0         │
#> │ Other Faiths    ┆ 20.0  ┆ 33.0    ┆ 40.0    ┆ … ┆ 46.0     ┆ 40.0      ┆ 41.0  ┆ 71.0         │
#> │ Other World     ┆ 5.0   ┆ 2.0     ┆ 3.0     ┆ … ┆ 3.0      ┆ 4.0       ┆ 4.0   ┆ 8.0          │
#> │ Religions       ┆       ┆         ┆         ┆   ┆          ┆           ┆       ┆              │
#> │ Unaffiliated    ┆ 217.0 ┆ 299.0   ┆ 374.0   ┆ … ┆ 407.0    ┆ 321.0     ┆ 258.0 ┆ 597.0        │
#> └─────────────────┴───────┴─────────┴─────────┴───┴──────────┴───────────┴───────┴──────────────┘

pl_relig_income |>
  pl_pivot_longer(!religion, names_to = "income", values_to = "count")
#> shape: (180, 3)
#> ┌──────────────┬────────────────────┬───────┐
#> │ religion     ┆ income             ┆ count │
#> │ ---          ┆ ---                ┆ ---   │
#> │ str          ┆ str                ┆ f64   │
#> ╞══════════════╪════════════════════╪═══════╡
#> │ Agnostic     ┆ <$10k              ┆ 27.0  │
#> │ Agnostic     ┆ $10-20k            ┆ 34.0  │
#> │ Agnostic     ┆ $20-30k            ┆ 60.0  │
#> │ Agnostic     ┆ $30-40k            ┆ 81.0  │
#> │ …            ┆ …                  ┆ …     │
#> │ Unaffiliated ┆ $75-100k           ┆ 407.0 │
#> │ Unaffiliated ┆ $100-150k          ┆ 321.0 │
#> │ Unaffiliated ┆ >150k              ┆ 258.0 │
#> │ Unaffiliated ┆ Don't know/refused ┆ 597.0 │
#> └──────────────┴────────────────────┴───────┘