Change column order

Use relocate() to change column positions, using the same syntax as select() to make it easy to move blocks of columns at once.

Usage

# S3 method for class 'polars_data_frame'
relocate(.data, ..., .before = NULL, .after = NULL)

# S3 method for class 'polars_lazy_frame'
relocate(.data, ..., .before = NULL, .after = NULL)

Arguments

.data

A Polars Data/LazyFrame

...

Any expression accepted by dplyr::select(): variable names, column numbers, select helpers, etc.

.before, .after

Column name (either quoted or unquoted) that indicates the destination of columns selected by .... Supplying neither will move columns to the left-hand side; specifying both is an error.

Examples

dat <- as_polars_df(mtcars)

dat |>
  relocate(hp, vs, .before = cyl)
#> shape: (32, 11)
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ mpg  в”Ҷ hp    в”Ҷ vs  в”Ҷ cyl в”Ҷ вҖҰ в”Ҷ qsec  в”Ҷ am  в”Ҷ gear в”Ҷ carb в”Ӯ
#> в”Ӯ ---  в”Ҷ ---   в”Ҷ --- в”Ҷ --- в”Ҷ   в”Ҷ ---   в”Ҷ --- в”Ҷ ---  в”Ҷ ---  в”Ӯ
#> в”Ӯ f64  в”Ҷ f64   в”Ҷ f64 в”Ҷ f64 в”Ҷ   в”Ҷ f64   в”Ҷ f64 в”Ҷ f64  в”Ҷ f64  в”Ӯ
#> в•һв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Ў
#> в”Ӯ 21.0 в”Ҷ 110.0 в”Ҷ 0.0 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 16.46 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 21.0 в”Ҷ 110.0 в”Ҷ 0.0 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 17.02 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 22.8 в”Ҷ 93.0  в”Ҷ 1.0 в”Ҷ 4.0 в”Ҷ вҖҰ в”Ҷ 18.61 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 110.0 в”Ҷ 1.0 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 19.44 в”Ҷ 0.0 в”Ҷ 3.0  в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 18.7 в”Ҷ 175.0 в”Ҷ 0.0 в”Ҷ 8.0 в”Ҷ вҖҰ в”Ҷ 17.02 в”Ҷ 0.0 в”Ҷ 3.0  в”Ҷ 2.0  в”Ӯ
#> в”Ӯ вҖҰ    в”Ҷ вҖҰ     в”Ҷ вҖҰ   в”Ҷ вҖҰ   в”Ҷ вҖҰ в”Ҷ вҖҰ     в”Ҷ вҖҰ   в”Ҷ вҖҰ    в”Ҷ вҖҰ    в”Ӯ
#> в”Ӯ 30.4 в”Ҷ 113.0 в”Ҷ 1.0 в”Ҷ 4.0 в”Ҷ вҖҰ в”Ҷ 16.9  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 2.0  в”Ӯ
#> в”Ӯ 15.8 в”Ҷ 264.0 в”Ҷ 0.0 в”Ҷ 8.0 в”Ҷ вҖҰ в”Ҷ 14.5  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 19.7 в”Ҷ 175.0 в”Ҷ 0.0 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 15.5  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 6.0  в”Ӯ
#> в”Ӯ 15.0 в”Ҷ 335.0 в”Ҷ 0.0 в”Ҷ 8.0 в”Ҷ вҖҰ в”Ҷ 14.6  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 8.0  в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 109.0 в”Ҷ 1.0 в”Ҷ 4.0 в”Ҷ вҖҰ в”Ҷ 18.6  в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 2.0  в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

# if .before and .after are not specified, selected columns are moved to the
# first positions
dat |>
  relocate(hp, vs)
#> shape: (32, 11)
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ hp    в”Ҷ vs  в”Ҷ mpg  в”Ҷ cyl в”Ҷ вҖҰ в”Ҷ qsec  в”Ҷ am  в”Ҷ gear в”Ҷ carb в”Ӯ
#> в”Ӯ ---   в”Ҷ --- в”Ҷ ---  в”Ҷ --- в”Ҷ   в”Ҷ ---   в”Ҷ --- в”Ҷ ---  в”Ҷ ---  в”Ӯ
#> в”Ӯ f64   в”Ҷ f64 в”Ҷ f64  в”Ҷ f64 в”Ҷ   в”Ҷ f64   в”Ҷ f64 в”Ҷ f64  в”Ҷ f64  в”Ӯ
#> в•һв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Ў
#> в”Ӯ 110.0 в”Ҷ 0.0 в”Ҷ 21.0 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 16.46 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 110.0 в”Ҷ 0.0 в”Ҷ 21.0 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 17.02 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 93.0  в”Ҷ 1.0 в”Ҷ 22.8 в”Ҷ 4.0 в”Ҷ вҖҰ в”Ҷ 18.61 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 110.0 в”Ҷ 1.0 в”Ҷ 21.4 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 19.44 в”Ҷ 0.0 в”Ҷ 3.0  в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 175.0 в”Ҷ 0.0 в”Ҷ 18.7 в”Ҷ 8.0 в”Ҷ вҖҰ в”Ҷ 17.02 в”Ҷ 0.0 в”Ҷ 3.0  в”Ҷ 2.0  в”Ӯ
#> в”Ӯ вҖҰ     в”Ҷ вҖҰ   в”Ҷ вҖҰ    в”Ҷ вҖҰ   в”Ҷ вҖҰ в”Ҷ вҖҰ     в”Ҷ вҖҰ   в”Ҷ вҖҰ    в”Ҷ вҖҰ    в”Ӯ
#> в”Ӯ 113.0 в”Ҷ 1.0 в”Ҷ 30.4 в”Ҷ 4.0 в”Ҷ вҖҰ в”Ҷ 16.9  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 2.0  в”Ӯ
#> в”Ӯ 264.0 в”Ҷ 0.0 в”Ҷ 15.8 в”Ҷ 8.0 в”Ҷ вҖҰ в”Ҷ 14.5  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 175.0 в”Ҷ 0.0 в”Ҷ 19.7 в”Ҷ 6.0 в”Ҷ вҖҰ в”Ҷ 15.5  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 6.0  в”Ӯ
#> в”Ӯ 335.0 в”Ҷ 0.0 в”Ҷ 15.0 в”Ҷ 8.0 в”Ҷ вҖҰ в”Ҷ 14.6  в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 8.0  в”Ӯ
#> в”Ӯ 109.0 в”Ҷ 1.0 в”Ҷ 21.4 в”Ҷ 4.0 в”Ҷ вҖҰ в”Ҷ 18.6  в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 2.0  в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

# .before and .after can be quoted or unquoted
dat |>
  relocate(hp, vs, .after = "gear")
#> shape: (32, 11)
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ mpg  в”Ҷ cyl в”Ҷ disp  в”Ҷ drat в”Ҷ вҖҰ в”Ҷ gear в”Ҷ hp    в”Ҷ vs  в”Ҷ carb в”Ӯ
#> в”Ӯ ---  в”Ҷ --- в”Ҷ ---   в”Ҷ ---  в”Ҷ   в”Ҷ ---  в”Ҷ ---   в”Ҷ --- в”Ҷ ---  в”Ӯ
#> в”Ӯ f64  в”Ҷ f64 в”Ҷ f64   в”Ҷ f64  в”Ҷ   в”Ҷ f64  в”Ҷ f64   в”Ҷ f64 в”Ҷ f64  в”Ӯ
#> в•һв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Ў
#> в”Ӯ 21.0 в”Ҷ 6.0 в”Ҷ 160.0 в”Ҷ 3.9  в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 110.0 в”Ҷ 0.0 в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 21.0 в”Ҷ 6.0 в”Ҷ 160.0 в”Ҷ 3.9  в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 110.0 в”Ҷ 0.0 в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 22.8 в”Ҷ 4.0 в”Ҷ 108.0 в”Ҷ 3.85 в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 93.0  в”Ҷ 1.0 в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 6.0 в”Ҷ 258.0 в”Ҷ 3.08 в”Ҷ вҖҰ в”Ҷ 3.0  в”Ҷ 110.0 в”Ҷ 1.0 в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 18.7 в”Ҷ 8.0 в”Ҷ 360.0 в”Ҷ 3.15 в”Ҷ вҖҰ в”Ҷ 3.0  в”Ҷ 175.0 в”Ҷ 0.0 в”Ҷ 2.0  в”Ӯ
#> в”Ӯ вҖҰ    в”Ҷ вҖҰ   в”Ҷ вҖҰ     в”Ҷ вҖҰ    в”Ҷ вҖҰ в”Ҷ вҖҰ    в”Ҷ вҖҰ     в”Ҷ вҖҰ   в”Ҷ вҖҰ    в”Ӯ
#> в”Ӯ 30.4 в”Ҷ 4.0 в”Ҷ 95.1  в”Ҷ 3.77 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 113.0 в”Ҷ 1.0 в”Ҷ 2.0  в”Ӯ
#> в”Ӯ 15.8 в”Ҷ 8.0 в”Ҷ 351.0 в”Ҷ 4.22 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 264.0 в”Ҷ 0.0 в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 19.7 в”Ҷ 6.0 в”Ҷ 145.0 в”Ҷ 3.62 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 175.0 в”Ҷ 0.0 в”Ҷ 6.0  в”Ӯ
#> в”Ӯ 15.0 в”Ҷ 8.0 в”Ҷ 301.0 в”Ҷ 3.54 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 335.0 в”Ҷ 0.0 в”Ҷ 8.0  в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 4.0 в”Ҷ 121.0 в”Ҷ 4.11 в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 109.0 в”Ҷ 1.0 в”Ҷ 2.0  в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

# select helpers are also available
dat |>
  relocate(contains("[aeiou]"))
#> shape: (32, 11)
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ mpg  в”Ҷ cyl в”Ҷ disp  в”Ҷ hp    в”Ҷ вҖҰ в”Ҷ vs  в”Ҷ am  в”Ҷ gear в”Ҷ carb в”Ӯ
#> в”Ӯ ---  в”Ҷ --- в”Ҷ ---   в”Ҷ ---   в”Ҷ   в”Ҷ --- в”Ҷ --- в”Ҷ ---  в”Ҷ ---  в”Ӯ
#> в”Ӯ f64  в”Ҷ f64 в”Ҷ f64   в”Ҷ f64   в”Ҷ   в”Ҷ f64 в”Ҷ f64 в”Ҷ f64  в”Ҷ f64  в”Ӯ
#> в•һв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Ў
#> в”Ӯ 21.0 в”Ҷ 6.0 в”Ҷ 160.0 в”Ҷ 110.0 в”Ҷ вҖҰ в”Ҷ 0.0 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 21.0 в”Ҷ 6.0 в”Ҷ 160.0 в”Ҷ 110.0 в”Ҷ вҖҰ в”Ҷ 0.0 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 22.8 в”Ҷ 4.0 в”Ҷ 108.0 в”Ҷ 93.0  в”Ҷ вҖҰ в”Ҷ 1.0 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 6.0 в”Ҷ 258.0 в”Ҷ 110.0 в”Ҷ вҖҰ в”Ҷ 1.0 в”Ҷ 0.0 в”Ҷ 3.0  в”Ҷ 1.0  в”Ӯ
#> в”Ӯ 18.7 в”Ҷ 8.0 в”Ҷ 360.0 в”Ҷ 175.0 в”Ҷ вҖҰ в”Ҷ 0.0 в”Ҷ 0.0 в”Ҷ 3.0  в”Ҷ 2.0  в”Ӯ
#> в”Ӯ вҖҰ    в”Ҷ вҖҰ   в”Ҷ вҖҰ     в”Ҷ вҖҰ     в”Ҷ вҖҰ в”Ҷ вҖҰ   в”Ҷ вҖҰ   в”Ҷ вҖҰ    в”Ҷ вҖҰ    в”Ӯ
#> в”Ӯ 30.4 в”Ҷ 4.0 в”Ҷ 95.1  в”Ҷ 113.0 в”Ҷ вҖҰ в”Ҷ 1.0 в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 2.0  в”Ӯ
#> в”Ӯ 15.8 в”Ҷ 8.0 в”Ҷ 351.0 в”Ҷ 264.0 в”Ҷ вҖҰ в”Ҷ 0.0 в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 4.0  в”Ӯ
#> в”Ӯ 19.7 в”Ҷ 6.0 в”Ҷ 145.0 в”Ҷ 175.0 в”Ҷ вҖҰ в”Ҷ 0.0 в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 6.0  в”Ӯ
#> в”Ӯ 15.0 в”Ҷ 8.0 в”Ҷ 301.0 в”Ҷ 335.0 в”Ҷ вҖҰ в”Ҷ 0.0 в”Ҷ 1.0 в”Ҷ 5.0  в”Ҷ 8.0  в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 4.0 в”Ҷ 121.0 в”Ҷ 109.0 в”Ҷ вҖҰ в”Ҷ 1.0 в”Ҷ 1.0 в”Ҷ 4.0  в”Ҷ 2.0  в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

dat |>
  relocate(hp, vs, .after = last_col())
#> shape: (32, 11)
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ mpg  в”Ҷ cyl в”Ҷ disp  в”Ҷ drat в”Ҷ вҖҰ в”Ҷ gear в”Ҷ carb в”Ҷ hp    в”Ҷ vs  в”Ӯ
#> в”Ӯ ---  в”Ҷ --- в”Ҷ ---   в”Ҷ ---  в”Ҷ   в”Ҷ ---  в”Ҷ ---  в”Ҷ ---   в”Ҷ --- в”Ӯ
#> в”Ӯ f64  в”Ҷ f64 в”Ҷ f64   в”Ҷ f64  в”Ҷ   в”Ҷ f64  в”Ҷ f64  в”Ҷ f64   в”Ҷ f64 в”Ӯ
#> в•һв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•җв•җв•Әв•җв•җв•җв•җв•җв•Ў
#> в”Ӯ 21.0 в”Ҷ 6.0 в”Ҷ 160.0 в”Ҷ 3.9  в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 4.0  в”Ҷ 110.0 в”Ҷ 0.0 в”Ӯ
#> в”Ӯ 21.0 в”Ҷ 6.0 в”Ҷ 160.0 в”Ҷ 3.9  в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 4.0  в”Ҷ 110.0 в”Ҷ 0.0 в”Ӯ
#> в”Ӯ 22.8 в”Ҷ 4.0 в”Ҷ 108.0 в”Ҷ 3.85 в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 1.0  в”Ҷ 93.0  в”Ҷ 1.0 в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 6.0 в”Ҷ 258.0 в”Ҷ 3.08 в”Ҷ вҖҰ в”Ҷ 3.0  в”Ҷ 1.0  в”Ҷ 110.0 в”Ҷ 1.0 в”Ӯ
#> в”Ӯ 18.7 в”Ҷ 8.0 в”Ҷ 360.0 в”Ҷ 3.15 в”Ҷ вҖҰ в”Ҷ 3.0  в”Ҷ 2.0  в”Ҷ 175.0 в”Ҷ 0.0 в”Ӯ
#> в”Ӯ вҖҰ    в”Ҷ вҖҰ   в”Ҷ вҖҰ     в”Ҷ вҖҰ    в”Ҷ вҖҰ в”Ҷ вҖҰ    в”Ҷ вҖҰ    в”Ҷ вҖҰ     в”Ҷ вҖҰ   в”Ӯ
#> в”Ӯ 30.4 в”Ҷ 4.0 в”Ҷ 95.1  в”Ҷ 3.77 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 2.0  в”Ҷ 113.0 в”Ҷ 1.0 в”Ӯ
#> в”Ӯ 15.8 в”Ҷ 8.0 в”Ҷ 351.0 в”Ҷ 4.22 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 4.0  в”Ҷ 264.0 в”Ҷ 0.0 в”Ӯ
#> в”Ӯ 19.7 в”Ҷ 6.0 в”Ҷ 145.0 в”Ҷ 3.62 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 6.0  в”Ҷ 175.0 в”Ҷ 0.0 в”Ӯ
#> в”Ӯ 15.0 в”Ҷ 8.0 в”Ҷ 301.0 в”Ҷ 3.54 в”Ҷ вҖҰ в”Ҷ 5.0  в”Ҷ 8.0  в”Ҷ 335.0 в”Ҷ 0.0 в”Ӯ
#> в”Ӯ 21.4 в”Ҷ 4.0 в”Ҷ 121.0 в”Ҷ 4.11 в”Ҷ вҖҰ в”Ҷ 4.0  в”Ҷ 2.0  в”Ҷ 109.0 в”Ҷ 1.0 в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҳ