tidypolars: More Efficient Tidyverse Code, Using Polars in the Background

Polars is a cross-language tool for manipulating very large data. However, one drawback is that the R implementation has a syntax that will look odd to many R users who are not used to Python syntax. The objective of tidypolars is to improve the ease-of-use of Polars in R by providing tidyverse syntax to polars.

See also

Author

Maintainer: Etienne Bacher etienne.bacher@protonmail.com [copyright holder]

Other contributors:

Anatoly Tsyplenkov atsyplenkov@fastmail.com (ORCID) [contributor]

Yousa Mirage Yousa-Mirage@foxmail.com (ORCID) (Yousa-Mirage) [contributor]