Convert a Polars DataFrame to an R data.frame

This is a simple wrapper of $to_data_frame() present in polars. This is only to make this function work in a pipe workflow.

Usage

to_r(.data, shrink_i64 = TRUE)

Arguments

.data

A Polars DataFrame only (LazyFrames are not modified). Any other object is returned as-is.

shrink_i64

Try to shrink i64 to i32 or lower. See Details.

Details

Int64 is a format accepted in Polars but not natively in R (the package bit64 helps with that). Therefore, int64 values will give infinitely small values when converted to R. If shrink_i64 is TRUE, int64 columns will be shrunk to a lower int before being converted to R.

Examples

iris |>
  as_polars() |>
  filter(Sepal.Length > 6) |>
  to_r()
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1           7.0         3.2          4.7         1.4 versicolor
#> 2           6.4         3.2          4.5         1.5 versicolor
#> 3           6.9         3.1          4.9         1.5 versicolor
#> 4           6.5         2.8          4.6         1.5 versicolor
#> 5           6.3         3.3          4.7         1.6 versicolor
#> 6           6.6         2.9          4.6         1.3 versicolor
#> 7           6.1         2.9          4.7         1.4 versicolor
#> 8           6.7         3.1          4.4         1.4 versicolor
#> 9           6.2         2.2          4.5         1.5 versicolor
#> 10          6.1         2.8          4.0         1.3 versicolor
#> 11          6.3         2.5          4.9         1.5 versicolor
#> 12          6.1         2.8          4.7         1.2 versicolor
#> 13          6.4         2.9          4.3         1.3 versicolor
#> 14          6.6         3.0          4.4         1.4 versicolor
#> 15          6.8         2.8          4.8         1.4 versicolor
#> 16          6.7         3.0          5.0         1.7 versicolor
#> 17          6.7         3.1          4.7         1.5 versicolor
#> 18          6.3         2.3          4.4         1.3 versicolor
#> 19          6.1         3.0          4.6         1.4 versicolor
#> 20          6.2         2.9          4.3         1.3 versicolor
#> 21          6.3         3.3          6.0         2.5  virginica
#> 22          7.1         3.0          5.9         2.1  virginica
#> 23          6.3         2.9          5.6         1.8  virginica
#> 24          6.5         3.0          5.8         2.2  virginica
#> 25          7.6         3.0          6.6         2.1  virginica
#> 26          7.3         2.9          6.3         1.8  virginica
#> 27          6.7         2.5          5.8         1.8  virginica
#> 28          7.2         3.6          6.1         2.5  virginica
#> 29          6.5         3.2          5.1         2.0  virginica
#> 30          6.4         2.7          5.3         1.9  virginica
#> 31          6.8         3.0          5.5         2.1  virginica
#> 32          6.4         3.2          5.3         2.3  virginica
#> 33          6.5         3.0          5.5         1.8  virginica
#> 34          7.7         3.8          6.7         2.2  virginica
#> 35          7.7         2.6          6.9         2.3  virginica
#> 36          6.9         3.2          5.7         2.3  virginica
#> 37          7.7         2.8          6.7         2.0  virginica
#> 38          6.3         2.7          4.9         1.8  virginica
#> 39          6.7         3.3          5.7         2.1  virginica
#> 40          7.2         3.2          6.0         1.8  virginica
#> 41          6.2         2.8          4.8         1.8  virginica
#> 42          6.1         3.0          4.9         1.8  virginica
#> 43          6.4         2.8          5.6         2.1  virginica
#> 44          7.2         3.0          5.8         1.6  virginica
#> 45          7.4         2.8          6.1         1.9  virginica
#> 46          7.9         3.8          6.4         2.0  virginica
#> 47          6.4         2.8          5.6         2.2  virginica
#> 48          6.3         2.8          5.1         1.5  virginica
#> 49          6.1         2.6          5.6         1.4  virginica
#> 50          7.7         3.0          6.1         2.3  virginica
#> 51          6.3         3.4          5.6         2.4  virginica
#> 52          6.4         3.1          5.5         1.8  virginica
#> 53          6.9         3.1          5.4         2.1  virginica
#> 54          6.7         3.1          5.6         2.4  virginica
#> 55          6.9         3.1          5.1         2.3  virginica
#> 56          6.8         3.2          5.9         2.3  virginica
#> 57          6.7         3.3          5.7         2.5  virginica
#> 58          6.7         3.0          5.2         2.3  virginica
#> 59          6.3         2.5          5.0         1.9  virginica
#> 60          6.5         3.0          5.2         2.0  virginica
#> 61          6.2         3.4          5.4         2.3  virginica